requirements.txt
asgiref==3.2.3
Django==3.0.2
djangorestframework==3.11.0
pytz==2019.3
PyYAML==5.2
sqlparse==0.3.0
